requirements.txt
# FastAPI y ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# AI Providers
google-generativeai==0.3.2
anthropic==0.18.1

# Twilio
twilio==8.11.1

# HTTP Client
httpx==0.26.0

# Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
